feat: add handlers for changing directory, current directory
